Calculating height and width
Added code to automatically calculate the width and height of a scaled image when one of maxWidth or maxHeight is provided, but not both (this also corrects a bug where the image size is being increased when the maxWidth or maxHeight property is larger than the image dimensions and the image). Also using the EXIF information to correctly rotate the image.